client code refactoring and new endpoints
